+++ test/data-ingest/key.avsc
@@ -1,0 +1,7 @@
+{
+    "type": "record",
+    "name": "Key",
+    "fields": [
+        {"name": "key1", "type": "string"}
+    ]
+}
