+++ test/data-ingest/mzcompose.py
@@ -19,7 +19,7 @@ from materialize.checks.checks import Check
 from materialize.checks.executors import MzcomposeExecutor
 from materialize.checks.scenarios import *  # noqa: F401 F403
 from materialize.checks.scenarios import Scenario
-from materialize.mzcompose import Composition, WorkflowArgumentParser
+from materialize.mzcompose import Composition, WorkflowArgumentParser, Service
 from materialize.mzcompose.services import (
     Clusterd,
     SchemaRegistry,
@@ -34,11 +34,6 @@ from materialize.version_list import VersionsFromGit
 from materialize.checks.common import KAFKA_SCHEMA_WITH_SINGLE_STRING_FIELD
 from time import sleep
 
-from confluent_kafka import Producer
-from confluent_kafka.serialization import StringSerializer, SerializationContext, MessageField
-from confluent_kafka.schema_registry import SchemaRegistryClient
-from confluent_kafka.schema_registry.avro import AvroSerializer
-
 SERVICES = [
     Postgres(),
     Zookeeper(),
@@ -48,35 +43,8 @@ SERVICES = [
     Materialized(),
     TestdriveService(no_reset=True),
     Clusterd(name="clusterd1", options=["--scratch-directory=/mzdata/source_data"]),
+    Service("data-ingest", {"mzbuild": "data-ingest"}),
 ]
-
-def delivery_report(err, msg):
-    """
-    Reports the failure or success of a message delivery.
-
-    Args:
-        err (KafkaError): The error that occurred on None on success.
-
-        msg (Message): The message that was produced or failed.
-
-    Note:
-        In the delivery report callback the Message.key() and Message.value()
-        will be the binary format as encoded by any configured Serializers and
-        not the same object that was passed to produce().
-        If you wish to pass the original object(s) for key and value to delivery
-        report callback we recommend a bound callback or lambda where you pass
-        the objects along.
-    """
-
-    if err is not None:
-        print("Delivery failed for User record {}: {}".format(msg.key(), err))
-        return
-    print('User record {} successfully produced to {} [{}] at offset {}'.format(
-        msg.key(), msg.topic(), msg.partition(), msg.offset()))
-
-def idfn(d, ctx):
-    return d
-
 
 def workflow_default(c: Composition, parser: WorkflowArgumentParser) -> None:
     parser.add_argument(
@@ -118,6 +86,9 @@ def workflow_default(c: Composition, parser: WorkflowArgumentParser) -> None:
 
         $ kafka-create-topic topic=upsert-insert
 
+        $ kafka-ingest format=avro key-format=avro topic=upsert-insert key-schema=${keyschema} schema=${schema} repeat=1
+        {"key1": "A${kafka-ingest.iteration}"} {"f1": "A${kafka-ingest.iteration}"}
+
         > CREATE CONNECTION IF NOT EXISTS kafka_conn FOR KAFKA BROKER '${testdrive.kafka-addr}';
 
         > CREATE CONNECTION IF NOT EXISTS csr_conn FOR CONFLUENT SCHEMA REGISTRY URL '${testdrive.schema-registry-url}';
@@ -131,36 +102,6 @@ def workflow_default(c: Composition, parser: WorkflowArgumentParser) -> None:
         """
     ))
 
-    with open(f"user.avsc") as f:
-        schema_str = f.read()
-    
-    with open(f"key.avsc") as f:
-        key_schema_str = f.read()
-    
-    # docker port data-ingest-schema-registry-1 8081
-    schema_registry_conf = {'url': "http://localhost:32874"}
-    schema_registry_client = SchemaRegistryClient(schema_registry_conf)
-    
-    avro_serializer = AvroSerializer(schema_registry_client,
-                                     schema_str,
-                                     idfn)
-    
-    key_avro_serializer = AvroSerializer(schema_registry_client,
-                                         key_schema_str,
-                                         idfn)
-    
-    ## docker port data-ingest-kafka-1 9092
-    producer_conf = {'bootstrap.servers': "localhost:32873"}
-    producer = Producer(producer_conf)
-    
-    producer.poll(0.0)
-    topic = "testdrive-upsert-insert-2018493381"
-    producer.produce(topic=topic,
-                     key=key_avro_serializer({"key1": "A10000"}, SerializationContext(topic, MessageField.KEY)),
-                    value=avro_serializer({"f1": "A10000"}, SerializationContext(topic, MessageField.VALUE)),
-                    on_delivery=delivery_report)
-    producer.flush()
-
-    sleep(360000)
+    c.run("data-ingest")
 
     c.down(destroy_volumes=True)
