+++ test/data-ingest/user.avsc
@@ -1,0 +1,7 @@
+{
+    "type" : "record",
+    "name" : "test",
+    "fields" : [
+        {"name":"f1", "type":"string"}
+    ]
+}
